Fixing Tiles targeting and mouse cursor behavior.
git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@6658 c06c8d41-db1a-0410-9941-cceddc491573

+++ crawl-ref/source/tilesdl.cc
@@ -580,9 +580,9 @@ int TilesFramework::getch_ck()
             }
         }
 
-        bool show_tooltip = 
-            ((ticks - m_last_tick_moved > Options.tile_tooltip_ms)
-                && ticks > m_last_tick_moved);
+        bool show_tooltip = ((ticks - m_last_tick_moved 
+                                > (unsigned int)Options.tile_tooltip_ms)
+                                && ticks > m_last_tick_moved);
 
         if (show_tooltip)
         {
